Add Delete ApexLogs button in popup org tab
